# CUDA: attention sinks for mma FlashAttention