This PR adds support for attention sinks to the CUDA FlashAttention kernel using tensor cores. On Turing or newer attention sinks are now fully supported.

<details>
<summary>Performance changes</summary>

| GPU         | Model                  |     Microbatch size | Test      |     t/s master |     t/s cuda-fa-mma-sink-2 |     Speedup |
| :---------  | :--------------------- | ------------------: | :-------- | -------------: | -------------------------: | ----------: |
| RTX 3090    | gpt-oss 20B MXFP4 MoE  |                   1 | pp16384   |         140.73 |                     157.42 |        1.12 |
| RTX 3090    | gpt-oss 20B MXFP4 MoE  |                   2 | pp16384   |         127.88 |                     144.44 |        1.13 |
| RTX 3090    | gpt-oss 20B MXFP4 MoE  |                   4 | pp16384   |         189.83 |                     234.50 |        1.24 |
| RTX 3090    | gpt-oss 20B MXFP4 MoE  |                   8 | pp16384   |         254.20 |                     345.49 |        1.36 |
| RTX 3090    | gpt-oss 20B MXFP4 MoE  |                  16 | pp16384   |         322.00 |                     486.15 |        1.51 |
| RTX 3090    | gpt-oss 20B MXFP4 MoE  |                  32 | pp16384   |         380.96 |                     635.92 |        1.67 |
| RTX 3090    | gpt-oss 20B MXFP4 MoE  |                  64 | pp16384   |         504.90 |                    1076.35 |        2.13 |
| RTX 3090    | gpt-oss 20B MXFP4 MoE  |                 128 | pp16384   |         619.46 |                    1816.28 |        2.93 |
| RTX 3090    | gpt-oss 20B MXFP4 MoE  |                 256 | pp16384   |         676.09 |                    2622.94 |        3.88 |
| RTX 3090    | gpt-oss 20B MXFP4 MoE  |                 512 | pp16384   |         690.35 |                    3337.20 |        4.83 |
| RTX 3090    | gpt-oss 20B MXFP4 MoE  |                1024 | pp16384   |         708.09 |                    3888.88 |        5.49 |
| RTX 3090    | gpt-oss 20B MXFP4 MoE  |                2048 | pp16384   |         681.68 |                    4099.78 |        6.01 |
| RTX 3090    | gpt-oss 20B MXFP4 MoE  |                4096 | pp16384   |         682.19 |                    4106.01 |        6.02 |
| RTX 3090    | gpt-oss 20B MXFP4 MoE  |                8192 | pp16384   |         682.25 |                    4101.10 |        6.01 |
| RTX 3090    | gpt-oss 20B MXFP4 MoE  |               16384 | pp16384   |         681.61 |                    4086.37 |        6.00 |
| RTX 4090    | gpt-oss 20B MXFP4 MoE  |                   1 | pp16384   |         238.95 |                     244.15 |        1.02 |
| RTX 4090    | gpt-oss 20B MXFP4 MoE  |                   2 | pp16384   |         199.57 |                     208.26 |        1.04 |
| RTX 4090    | gpt-oss 20B MXFP4 MoE  |                   4 | pp16384   |         322.24 |                     358.46 |        1.11 |
| RTX 4090    | gpt-oss 20B MXFP4 MoE  |                   8 | pp16384   |         476.46 |                     566.93 |        1.19 |
| RTX 4090    | gpt-oss 20B MXFP4 MoE  |                  16 | pp16384   |         641.87 |                     840.95 |        1.31 |
| RTX 4090    | gpt-oss 20B MXFP4 MoE  |                  32 | pp16384   |         818.55 |                    1190.88 |        1.45 |
| RTX 4090    | gpt-oss 20B MXFP4 MoE  |                  64 | pp16384   |        1138.38 |                    2024.07 |        1.78 |
| RTX 4090    | gpt-oss 20B MXFP4 MoE  |                 128 | pp16384   |        1477.22 |                    3453.34 |        2.34 |
| RTX 4090    | gpt-oss 20B MXFP4 MoE  |                 256 | pp16384   |        1650.90 |                    4900.11 |        2.97 |
| RTX 4090    | gpt-oss 20B MXFP4 MoE  |                 512 | pp16384   |        1704.01 |                    6090.15 |        3.57 |
| RTX 4090    | gpt-oss 20B MXFP4 MoE  |                1024 | pp16384   |        1730.74 |                    6650.04 |        3.84 |
| RTX 4090    | gpt-oss 20B MXFP4 MoE  |                2048 | pp16384   |        1651.59 |                    6596.91 |        3.99 |
| RTX 4090    | gpt-oss 20B MXFP4 MoE  |                4096 | pp16384   |        1651.71 |                    6582.91 |        3.99 |
| RTX 4090    | gpt-oss 20B MXFP4 MoE  |                8192 | pp16384   |        1648.47 |                    6581.47 |        3.99 |
| RTX 4090    | gpt-oss 20B MXFP4 MoE  |               16384 | pp16384   |        1648.81 |                    6570.34 |        3.98 |
| 3x RTX 4090 | gpt-oss 120B MXFP4 MoE |                   1 | pp16384   |         217.68 |                     221.46 |        1.02 |
| 3x RTX 4090 | gpt-oss 120B MXFP4 MoE |                   2 | pp16384   |         115.67 |                     120.47 |        1.04 |
| 3x RTX 4090 | gpt-oss 120B MXFP4 MoE |                   4 | pp16384   |         178.51 |                     196.02 |        1.10 |
| 3x RTX 4090 | gpt-oss 120B MXFP4 MoE |                   8 | pp16384   |         252.54 |                     291.57 |        1.15 |
| 3x RTX 4090 | gpt-oss 120B MXFP4 MoE |                  16 | pp16384   |         324.61 |                     397.50 |        1.22 |
| 3x RTX 4090 | gpt-oss 120B MXFP4 MoE |                  32 | pp16384   |         389.73 |                     508.91 |        1.31 |
| 3x RTX 4090 | gpt-oss 120B MXFP4 MoE |                  64 | pp16384   |         541.84 |                     795.13 |        1.47 |
| 3x RTX 4090 | gpt-oss 120B MXFP4 MoE |                 128 | pp16384   |         705.35 |                    1217.79 |        1.73 |
| 3x RTX 4090 | gpt-oss 120B MXFP4 MoE |                 256 | pp16384   |         867.21 |                    1811.41 |        2.09 |
| 3x RTX 4090 | gpt-oss 120B MXFP4 MoE |                 512 | pp16384   |         967.18 |                    2426.25 |        2.51 |
| 3x RTX 4090 | gpt-oss 120B MXFP4 MoE |                1024 | pp16384   |        1028.36 |                    3076.55 |        2.99 |
| 3x RTX 4090 | gpt-oss 120B MXFP4 MoE |                2048 | pp16384   |        1007.56 |                    3343.48 |        3.32 |
| 3x RTX 4090 | gpt-oss 120B MXFP4 MoE |                4096 | pp16384   |        1006.97 |                    3339.11 |        3.32 |
| 3x RTX 4090 | gpt-oss 120B MXFP4 MoE |                8192 | pp16384   |        1006.49 |                    3315.28 |        3.29 |
| 3x RTX 4090 | gpt-oss 120B MXFP4 MoE |               16384 | pp16384   |        1006.69 |                    3335.57 |        3.31 |

</details>