@slaren 2025-08-07
The reason you get the same performance for ubatch sizes >= 2048 is likely because the default batch size of `llama-bench` is 2048, and you haven't changed it. You need to increase the batch size as well if you want to try ubatch sizes bigger than 2048, otherwise llama-bench will only submit batches of 2048 tokens to llama.cpp.