@slaren 2025-08-07
It might have been a fluke, I regenerated the same prompt a few times and it didn't happen again.