@JohannesGaessler 2025-08-07
When I set `SOFTMAX_FTZ_THRESHOLD` to -50 instead of -20 the perplexity on 1 chunk of wikitext goes down from 214.9317 to 191.5787. So maybe this is what's causing the issues since the vector kernels don't use it. This was kind of a hack in the first place to avoid NaNs from FP16 underflows, if it's now causing issues with GPT-OSS there may be a better way to handle this.